+++ primitives/chain-millau/src/lib.rs
@@ -26,7 +26,7 @@ use bp_messages::{
 	ChainWithMessages, InboundMessageDetails, LaneId, MessageNonce, MessagePayload,
 	OutboundMessageDetails,
 };
-use bp_runtime::{decl_bridge_runtime_apis, Chain};
+use bp_runtime::{decl_bridge_runtime_apis, Chain, ChainId};
 use frame_support::{
 	dispatch::DispatchClass,
 	weights::{constants::WEIGHT_REF_TIME_PER_SECOND, IdentityFee, Weight},
@@ -157,6 +157,8 @@ pub type WeightToFee = IdentityFee<Balance>;
 pub struct Millau;
 
 impl Chain for Millau {
+	const ID: ChainId = bp_runtime::MILLAU_CHAIN_ID;
+
 	type BlockNumber = BlockNumber;
 	type Hash = Hash;
 	type Hasher = Hasher;
