+++ primitives/chain-rialto-parachain/src/lib.rs
@@ -22,7 +22,7 @@ use bp_messages::{
 	ChainWithMessages, InboundMessageDetails, LaneId, MessageNonce, MessagePayload,
 	OutboundMessageDetails,
 };
-use bp_runtime::{decl_bridge_runtime_apis, Chain, Parachain};
+use bp_runtime::{decl_bridge_runtime_apis, Chain, ChainId, Parachain};
 use frame_support::{
 	dispatch::DispatchClass,
 	weights::{constants::WEIGHT_REF_TIME_PER_SECOND, IdentityFee, Weight},
@@ -108,6 +108,8 @@ pub type WeightToFee = IdentityFee<Balance>;
 pub struct RialtoParachain;
 
 impl Chain for RialtoParachain {
+	const ID: ChainId = bp_runtime::RIALTO_PARACHAIN_CHAIN_ID;
+
 	type BlockNumber = BlockNumber;
 	type Hash = Hash;
 	type Hasher = Hasher;
