+++ relays/client-bridge-hub-kusama/Cargo.toml
@@ -16,7 +16,6 @@ bp-bridge-hub-kusama = { path = "../../primitives/chain-bridge-hub-kusama" }
 bp-bridge-hub-polkadot = { path = "../../primitives/chain-bridge-hub-polkadot" }
 bp-header-chain = { path = "../../primitives/header-chain" }
 bp-parachains = { path = "../../primitives/parachains" }
-bp-runtime = { path = "../../primitives/runtime" }
 bp-polkadot = { path = "../../primitives/chain-polkadot" }
 
 bridge-runtime-common = { path = "../../bin/runtime-common" }
