+++ relays/client-bridge-hub-polkadot/src/lib.rs
@@ -17,7 +17,6 @@
 //! Types used to connect to the BridgeHub-Polkadot-Substrate parachain.
 
 use bp_bridge_hub_polkadot::{BridgeHubSignedExtension, AVERAGE_BLOCK_INTERVAL};
-use bp_runtime::ChainId;
 use codec::Encode;
 use relay_substrate_client::{
 	Chain, ChainWithBalances, ChainWithMessages, ChainWithTransactions, ChainWithUtilityPallet,
@@ -41,7 +40,6 @@ impl UnderlyingChainProvider for BridgeHubPolkadot {
 }
 
 impl Chain for BridgeHubPolkadot {
-	const ID: ChainId = bp_runtime::BRIDGE_HUB_POLKADOT_CHAIN_ID;
 	const NAME: &'static str = "BridgeHubPolkadot";
 	const BEST_FINALIZED_HEADER_ID_METHOD: &'static str =
 		bp_bridge_hub_polkadot::BEST_FINALIZED_BRIDGE_HUB_POLKADOT_HEADER_METHOD;
