+++ relays/client-bridge-hub-wococo/Cargo.toml
@@ -18,7 +18,6 @@ bp-header-chain = { path = "../../primitives/header-chain" }
 bp-parachains = { path = "../../primitives/parachains" }
 bp-polkadot-core = { path = "../../primitives/polkadot-core" }
 bp-rococo = { path = "../../primitives/chain-rococo" }
-bp-runtime = { path = "../../primitives/runtime" }
 
 bridge-runtime-common = { path = "../../bin/runtime-common" }
 relay-bridge-hub-rococo-client = { path = "../client-bridge-hub-rococo" }
