+++ relays/client-kusama/Cargo.toml
@@ -12,7 +12,6 @@ relay-utils = { path = "../utils" }
 # Bridge dependencies
 
 bp-kusama = { path = "../../primitives/chain-kusama" }
-bp-runtime = { path = "../../primitives/runtime" }
 
 # Substrate Dependencies
