+++ relays/client-millau/Cargo.toml
@@ -13,7 +13,6 @@ relay-utils = { path = "../utils" }
 # Supported Chains
 
 bp-millau = { path = "../../primitives/chain-millau" }
-bp-runtime = { path = "../../primitives/runtime" }
 millau-runtime = { path = "../../bin/millau/runtime" }
 
 # Substrate Dependencies
