+++ relays/client-polkadot/Cargo.toml
@@ -12,7 +12,6 @@ relay-utils = { path = "../utils" }
 # Bridge dependencies
 
 bp-polkadot = { path = "../../primitives/chain-polkadot" }
-bp-runtime = { path = "../../primitives/runtime" }
 
 # Substrate Dependencies
