+++ relays/client-rialto/Cargo.toml
@@ -13,7 +13,6 @@ relay-utils = { path = "../utils" }
 # Bridge dependencies
 
 bp-rialto = { path = "../../primitives/chain-rialto" }
-bp-runtime = { path = "../../primitives/runtime" }
 rialto-runtime = { path = "../../bin/rialto/runtime" }
 
 # Substrate Dependencies
