+++ relays/client-rococo/Cargo.toml
@@ -12,7 +12,6 @@ relay-utils = { path = "../utils" }
 # Bridge dependencies
 
 bp-rococo = { path = "../../primitives/chain-rococo" }
-bp-runtime = { path = "../../primitives/runtime" }
 
 # Substrate Dependencies
