+++ relays/client-westend/Cargo.toml
@@ -11,7 +11,6 @@ relay-utils = { path = "../utils" }
 
 # Bridge dependencies
 
-bp-runtime = { path = "../../primitives/runtime" }
 bp-westend = { path = "../../primitives/chain-westend" }
 
 # Substrate Dependencies
