+++ relays/client-wococo/src/lib.rs
@@ -16,7 +16,6 @@
 
 //! Types used to connect to the Wococo-Substrate chain.
 
-use bp_runtime::ChainId;
 use relay_substrate_client::{Chain, ChainWithBalances, RelayChain, UnderlyingChainProvider};
 use sp_core::storage::StorageKey;
 use std::time::Duration;
@@ -36,7 +35,6 @@ impl UnderlyingChainProvider for Wococo {
 }
 
 impl Chain for Wococo {
-	const ID: ChainId = bp_runtime::WOCOCO_CHAIN_ID;
 	const NAME: &'static str = "Wococo";
 	const BEST_FINALIZED_HEADER_ID_METHOD: &'static str =
 		bp_wococo::BEST_FINALIZED_WOCOCO_HEADER_METHOD;
